README
# ArduinoCommunication
The smart devices around us use a kind of communication. The multi-sensor communication system forms the Internet of Things (IoT). The project basically uses two types of communication. The types of communication it uses are wireless and bluetooth. Both communication types enables the device to send and receive packets without human intervention. Wireless communication is used to send packets to the database, while bluetooth is used to communicate between two devices. If the wireless communication fails, it means that there is no internet then device number 1 via bluetooth checks if the other device has internet to send data on its behalf. If there is no internet on both devices then they store their data on the SD card until the internet is back. The project contains several sensors which provide data about the device. The data is processed and compared with the second device. Data is collected in database.